+++ src/cli/types.ts
@@ -9,3 +9,8 @@ export type Repository = {
   fullName: string;
   description: string;
 };
+
+export type RepoShort = {
+  shortName: string;
+  description: string;
+};